+++ CHANGES.rst
@@ -7,6 +7,8 @@ Release 0.3.2.dev
 New Features:
 
 * automatically generated lists of figures and tables
+* paragraphs now provide default tab stops (proportional to font size) for
+  indentation
 * Sphinx frontend: the rinoh builder can be discovered by entry point
   (no more need to add 'rinoh.frontend.sphinx' to the list of extensions)
 * rinoh: set a return code of 1 when one or more referenced images could not be